# docs/swagger.yaml
basePath: /api/v1
definitions:
  github_com_Aliizi83_vohu_internal_ai_model.ToolCall:
    properties:
      arguments:
        additionalProperties: {}
        type: object
      id:
        type: string
      metadata:
        additionalProperties: {}
        type: object
      name:
        type: string
    type: object
  github_com_Aliizi83_vohu_internal_platform_conversation.CreateConversationRequest:
    properties:
      customModelId:
        description: |-
          CustomModelID, when set, names one of the caller's (or a global)
          custommodel.CustomModel presets to use for this conversation's
          credentials instead of the single per-user/global "openai" key —
          see Conversation.CustomModelID. Not validated here (conversation
          never imports custommodel); chat.buildLLM resolves and checks
          ownership lazily the first time the conversation is actually used.
        type: integer
      model:
        type: string
      provider:
        type: string
      title:
        maxLength: 255
        type: string
    required:
    - model
    - provider
    - title
    type: object
  github_com_Aliizi83_vohu_internal_platform_conversation.Response:
    properties:
      customModelId:
        type: integer
      id:
        type: integer
      model:
        type: string
      provider:
        type: string
      title:
        type: string
    type: object
  github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse:
    properties:
      error:
        type: string
      result: {}
      resultCode:
        $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ResultCode'
      success:
        type: boolean
      validationErrors:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.ValidationError'
        type: array
    type: object
  github_com_Aliizi83_vohu_internal_platform_shared.PagedList-github_com_Aliizi83_vohu_internal_platform_conversation_Response:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_conversation.Response'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_chat_MessageResponse:
    properties:
      hasNextPage:
        type: boolean
      hasPreviousPage:
        type: boolean
      items:
        items:
          $ref: '#/definitions/internal_platform_chat.MessageResponse'
        type: array
      pageNumber:
        type: integer
      pageSize:
        type: integer
      totalPages:
        type: integer
      totalRows:
        type: integer
    type: object
...
  github_com_Aliizi83_vohu_internal_platform_shared.ResultCode:
    enum:
    - 0
    - 40001
    - 40101
    - 40301
    - 40401
    - 40901
    - 50001
    type: integer
    x-enum-varnames:
    - ResultSuccess
    - ResultValidationError
    - ResultAuthError
    - ResultForbiddenError
    - ResultNotFoundError
    - ResultConflictError
    - ResultInternalError
  github_com_Aliizi83_vohu_internal_platform_shared.ValidationError:
    properties:
      message:
        type: string
      property:
        type: string
      tag:
        type: string
      value:
        type: string
    type: object
...
  internal_platform_chat.MessageResponse:
    properties:
      content:
        type: string
      role:
        type: string
      toolCalls:
        items:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_ai_model.ToolCall'
        type: array
      toolResults:
        items:
          $ref: '#/definitions/internal_platform_chat.ToolResultResponse'
        type: array
    type: object
  internal_platform_chat.SendMessageRequest:
    properties:
      content:
        type: string
    required:
    - content
    type: object
  internal_platform_chat.ToolResultResponse:
    properties:
      error:
        type: string
      name:
        type: string
      result: {}
      toolCallId:
        type: string
    type: object
...
info:
  contact: {}
  description: HTTP API for Vohu's multi-user platform — users, roles, polymorphic
    resource access, SSH connections, conversations/chat, provider keys, custom models,
    and the agent tool catalog. Every endpoint responds through shared.BaseResponse
    (see internal/platform/shared/response.go) — {result, success, resultCode, validationErrors,
    error} — regardless of success or failure.
  title: Vohu Platform API
  version: "1.0"
paths:
...
  /conversations:
    get:
      description: Lists the caller's own conversations. There's no separate access-policy
        gate here beyond being authenticated — conversations are scoped to their owner
        by construction.
      parameters:
      - description: Page number, default 1
        in: query
        name: pageNumber
        type: integer
      - description: Page size, default 10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.PagedList-github_com_Aliizi83_vohu_internal_platform_conversation_Response'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List my conversations
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: Starts a new conversation — provider and model are fixed for its
        whole lifetime once created. customModelId, if set, names one of the caller's
        (or a global) custom model presets to use instead of the account's single
        per-provider key.
      parameters:
      - description: New conversation
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_conversation.CreateConversationRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_conversation.Response'
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Create a conversation
      tags:
      - chat
  /conversations/{id}/messages:
    get:
      description: Returns one page of a conversation's messages, newest first (page
        1 is the most recent; higher page numbers reach further into the past) — for
        scroll-up-to-load-older-history. 404s for a conversation the caller doesn't
        own.
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Page number, default 1 (most recent)
        in: query
        name: pageNumber
        type: integer
      - description: Page size, default 10
        in: query
        name: pageSize
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
            - properties:
                result:
                  $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.PagedList-internal_platform_chat_MessageResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: List a conversation's messages
      tags:
      - chat
    post:
      consumes:
      - application/json
      description: 'Sends a message and streams the reply as a Server-Sent Events
        response (Content-Type: text/event-stream) — this is NOT a plain JSON endpoint
        despite the shared.BaseResponse envelope every other route uses; it''s documented
        here for completeness but tools like "Try it out" won''t render it usefully.
        Events, in order: zero or more "chunk" (data is a raw string — one piece of
        assistant text as it streams in), then either "done" (data is []MessageResponse
        — every message this turn produced: the assistant''s reply and any tool call/result
        pairs, already persisted) or "error" (data is a plain error string; nothing
        was persisted).'
      parameters:
      - description: Conversation ID
        in: path
        name: id
        required: true
        type: integer
      - description: Message content
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/internal_platform_chat.SendMessageRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: SSE stream — see description for event shapes
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/github_com_Aliizi83_vohu_internal_platform_shared.BaseResponse'
      security:
      - BearerAuth: []
      summary: Send a message (SSE stream)
      tags:
      - chat
...
securityDefinitions:
  BearerAuth:
    description: Type "Bearer" followed by a space and the access token from POST
      /auth/login or /auth/refresh.
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"